requirements.txt
Django==1.10.4
django-appconf==1.0.2
django-crispy-forms==1.6.1
django-embed-video==1.1.0
django-imagekit==3.3
django-registration-redux==1.4
Django-Select2==5.8.10
django-tinymce==2.4.0
html2text==2014.12.29
pilkit==1.1.13
Pillow==3.4.2
requests==2.12.4
six==1.10.0
